add nginx config example

+++ admin/admin-cookie-wall.php
@@ -166,8 +166,26 @@ class Admin_Cookie_Wall {
 		if( !is_dir( $plugin_admin_path . '/config_files' ) ) {
 			mkdir( $plugin_admin_path . '/config_files' );
 		}
+		$host = $_SERVER['HTTP_HOST'];
 
-		$content = 'volgt nog';
+		$content = '
+		set $ll_cookie_exist \'0\';
+		if ( $http_user_agent ~* \'(Internet\ Explorer|MSIE|Chrome|Safari|Firefox|Windows|Opera|iphone|ipad|andriod|blackberry)\' ) { 
+			set $ll_cookie_exist \'1\';
+		}
+		if ( $http_cookie ~ "ll_cookie_wall=ll_cookie_wall" ) { 
+			set $ll_cookie_exist \'0\'; 
+		}
+		if ($request_uri ~ ^/cookie_wall\?url_redirect ) {
+			set $ll_cookie_exist \'0\';
+		}
+		if ($request_uri ~ ^/wp-content ) {
+		    set $ll_cookie_exist \'0\';
+		}
+		if ( $ll_cookie_exist = \'1\' ) { 
+			return 302 http://'.$host.'/cookie_wall?url_redirect=$scheme://$host$request_uri; 
+		}
+		';
 
 		file_put_contents( $nginx_config_path, $content );
